fix: prefMenuをコンポーネント化
