+++ front/src/Components/Home.js
@@ -12,12 +12,11 @@ import TextField from 'material-ui/TextField';
 import Menu, { MenuItem } from 'material-ui/Menu';
 import Paper from 'material-ui/Paper';
 import Button from 'material-ui/Button';
-import InputLabel from 'material-ui/Input/InputLabel';
-import FormControl from 'material-ui/Form/FormControl';
-import List, { ListItem, ListItemText } from 'material-ui/List';
 
 import match from 'autosuggest-highlight/match';
 import parse from 'autosuggest-highlight/parse';
+
+import PrefMenu from './PrefMenu';
 
 import * as ErrorActions from '../Actions/Error';
 import * as AutosuggestActions from '../Actions/Autosuggest';
@@ -57,27 +56,12 @@ const styles = theme =>({
   },
 });
 
-const prefs = [
-  '選択してください','北海道', '青森県', '岩手県', '宮城県', '秋田県',
-    '山形県', '福島県', '茨城県', '栃木県', '群馬県',
-    '埼玉県', '千葉県', '東京都', '神奈川県', '新潟県',
-    '富山県', '石川県', '福井県', '山梨県', '長野県',
-    '岐阜県', '静岡県', '愛知県', '三重県', '滋賀県',
-    '京都府', '大阪府', '兵庫県', '奈良県', '和歌山県',
-    '鳥取県', '島根県', '岡山県', '広島県', '山口県',
-    '徳島県', '香川県', '愛媛県', '高知県', '福岡県',
-    '佐賀県', '長崎県', '熊本県', '大分県', '宮崎県',
-    '鹿児島県', '沖縄県'
-];
-
 class Home extends React.Component {
   constructor(props) {
     super(props);
     this.state={
       keyword: '',
       selectedPrefIndex: 0,
-      isPrefOpen: false,
-      anchorEl: undefined,
     };
     props.errorActions.removeError();
   }
@@ -175,54 +159,6 @@ class Home extends React.Component {
     });
   };
 
-
-  handleClickPrefListItem = event => {
-    this.setState({ isPrefOpen: true, anchorEl: event.currentTarget });
-  };
-
-  handleMenuItemClick = (event, index) => {
-    this.setState({ selectedPrefIndex: index, isPrefOpen: false });
-  };
-
-  handleRequestClose = () => {
-    this.setState({ isPrefOpen: false });
-  };
-
-  renderPrefSelection(){
-    return (
-      <div>
-        <FormControl style={{paddingTop: 20, width: '100%'}}>
-          <InputLabel htmlFor="pref-item">都道府県</InputLabel>
-          <List id="pref-item">
-            <ListItem
-              button
-              onClick={this.handleClickPrefListItem}
-              >
-              <ListItemText
-                primary={prefs[this.state.selectedPrefIndex]}
-              />
-            </ListItem>
-          </List>
-        </FormControl>
-        <Menu
-          anchorEl={this.state.anchorEl}
-          open={this.state.isPrefOpen}
-          onRequestClose={this.handleRequestClose}
-        >
-          {prefs.map((option, index) => (
-              <MenuItem
-                key={option}
-                selected={index === this.state.selectedIndex}
-                onClick={event => this.handleMenuItemClick(event, index)}
-              >
-                {option}
-              </MenuItem>
-            ))}
-        </Menu>
-      </div>
-    );
-  }
-
   handleSubmit = () => {
     const { eventsAction, history } = this.props;
     const { keyword, selectedPrefIndex } = this.state;
@@ -272,7 +208,7 @@ class Home extends React.Component {
                   onChange: this.handleKeywordChange,
                 }}
               />
-            {this.renderPrefSelection()}
+              <PrefMenu onSelectPref={(event,index) => { this.setState({selectedPrefIndex: index})}}/>
               <Button
                 raised
                 color="primary"
